Puzzles 8, 9 and 10 solved completely.

+++ java/src/com/s13g/aoc/Puzzle.java
@@ -29,7 +29,7 @@ public interface Puzzle {
       this.solutionB = solutionB;
     }
 
-    public Solution(int solutionA, int solutionB) {
+    public Solution(long solutionA, long solutionB) {
       this.solutionA = String.valueOf(solutionA);
       this.solutionB = String.valueOf(solutionB);
     }
